pyproject.toml
[project]
name = "kimi-go-brr"
version = "0.1.0"
description = "Self-improving SQL analytics agent hackathon project"
requires-python = ">=3.11"
dependencies = [
  "datasets>=4.1.1",
  "deepagents",
  "langchain-openai>=1.0.1",
  "langchain-litellm>=0.2.2",
  "litellm>=1.76.2",
  "openai>=2.0.0",
  "psycopg2-binary>=2.9.10",
  "python-dotenv>=1.1.1",
  "fastapi>=0.135.2",
  "uvicorn>=0.42.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "eval", "pipeline", "training"]
